fixed kubectl cache

+++ Jenkinsfile
@@ -4,57 +4,54 @@ pipeline {
     parameters {
         booleanParam(name: 'SKIP_DOCKER_BUILD', defaultValue: false, description: 'Skip Docker build steps')
     }
+
     environment {
         backendImageName = ''
         frontendImageName = ''
     }
 
     stages {
-        stage('Check main branch') {
-            when {
-                branch 'main'
-            }
+        stage('Determine Image Names') {
             steps {
                 script {
-                    backendImageName = '1md3nd/todo-backend-prod'
-                    frontendImageName = '1md3nd/todo-frontend-prod'
+                    if (env.BRANCH_NAME == 'main') {
+                        env.backendImageName = '1md3nd/todo-backend-prod'
+                        env.frontendImageName = '1md3nd/todo-frontend-prod'
+                    } else if (env.BRANCH_NAME.startsWith('dev-')) {
+                        env.backendImageName = '1md3nd/todo-backend-dev'
+                        env.frontendImageName = '1md3nd/todo-frontend-dev'
+                    }
                 }
             }
         }
-        stage('Check dev branch') {
-            when {
-                branch 'dev-*'
-            }
-            steps {
-                script {
-                    backendImageName = '1md3nd/todo-backend-dev'
-                    frontendImageName = '1md3nd/todo-frontend-dev'
-                }
-            }
-        }
-        stage('Build And Deploy Image') {
+
+        stage('Build And Deploy Images') {
             when {
                 expression { !params.SKIP_DOCKER_BUILD }
             }
             steps {
                 script {
-                    buildAndDeployImage('backend', backendImageName, './backend')
-                    buildAndDeployImage('frontend', frontendImageName, './frontend')
+                    buildAndDeployImage('backend', env.backendImageName, './backend')
+                    buildAndDeployImage('frontend', env.frontendImageName, './frontend')
                 }
             }
         }
+
         stage('Deploy on Kubernetes') {
             steps {
                 withKubeConfig([credentialsId: 'k8s']) {
-                    sh '''
-                    curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
-                    chmod +x kubectl
-                    mkdir -p ~/.local/bin
-                    mv ./kubectl ~/.local/bin/kubectl
-                    ~/.local/bin/kubectl apply -f k8s/misc
-                    ~/.local/bin/kubectl apply -f k8s/deployments
-                    ~/.local/bin/kubectl apply -f k8s/services
-                    '''
+                    script {
+                        if (!fileExists('~/.local/bin/kubectl')) {
+                            sh '''
+                                curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
+                                chmod +x kubectl
+                                mkdir -p ~/.local/bin
+                                mv ./kubectl ~/.local/bin/kubectl
+                            '''
+                        }
+
+                        sh "~/.local/bin/kubectl apply -f k8s"
+                    }
                 }
             }
         }
@@ -67,7 +64,7 @@ def buildAndDeployImage(imageType, imageName, path) {
         image = docker.build("${imageName}:${env.BUILD_ID}", "${path}")
     } catch (Exception e) {
         currentBuild.result = 'FAILURE'
-        echo "Failed to build ${imageType} image: ${e.message}"
+        error "Failed to build ${imageType} image: ${e.message}"
     } finally {
         echo "Done building ${imageType} image"
     }
@@ -79,7 +76,7 @@ def buildAndDeployImage(imageType, imageName, path) {
         }
     } catch (Exception e) {
         currentBuild.result = 'FAILURE'
-        echo "Failed to deploy ${imageType} image: ${e.message}"
+        error "Failed to deploy ${imageType} image: ${e.message}"
     } finally {
         echo "Done deploying ${imageType} image"
     }
